+++ blocklets/image-bin/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.11.119 (May 08, 2025)
+
+- refactor: polish uploader auto focus logic
+
 ## 0.11.118 (May 07, 2025)
 
 - chore: update deps