+++ packages/uploader/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.86 (May 08, 2025)
+
+- refactor: polish uploader auto focus logic
+
 ## 0.1.85 (May 07, 2025)
 
 - chore: update deps